pyproject.toml
[build-system]
requires = ["setuptools>=65.5.1", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-pride-archive-search"
version = "1.0.0"
description = "PRIDE Archive search service implementing  Model Context Protocol (MCP) tools"
readme = "README.md"
authors = [
    {name = "Suresh Hewapathirana", email = "sureshhewabi@gmail.com"}
]
license = "MIT"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastmcp",
    "rich",
    "pydantic",
    "python-dotenv",
]

[project.scripts]
mcp_pride_archive_search = "mcp_pride_archive_search.server:main"

[tool.setuptools]
packages = ["mcp_pride_archive_search",
            "mcp_pride_archive_search.models",
            "mcp_pride_archive_search.services",
            "mcp_pride_archive_search.utils"]